Extract MemberPage to its own file
This step is required so that scraper_test can test MemberPage data against a
yaml file. scraper_test needs to require the individual class.
